+++ fdr-plugin/src/test/java/com/dwarfeng/fdr/plugin/configuration/ConsumerConfiguration.java
@@ -1,0 +1,8 @@
+package com.dwarfeng.fdr.plugin.configuration;
+
+/**
+ * @author DwArFeng
+ * @since 1.2.0.a
+ */
+public class ConsumerConfiguration {
+}